agriconnect_guide/requirements.txt
asgiref==3.11.1
bleach==6.3.0
Django==6.0.3
django-axes==8.3.1
django-cors-headers==4.9.0
djangorestframework==3.17.1
packaging==26.0
pillow==12.2.0
python-decouple==3.8
setuptools==82.0.1
sqlparse==0.5.5
tzdata==2026.1
webencodings==0.5.1
wheel==0.46.3
whitenoise==6.12.0
